refactor: 상단바 중복 컴포넌트 제거 후 경로 수정
